Day 31 - Flashcards App/requirements.txt
customtkinter==4.5.9
pandas==1.4.3
pillow==9.2.0